// src/telegram-types.ts
export interface TelegramUser extends BaseUser {
  chat_id: string;
  user_name: string;
  language?: string;
  state?: string;
  created_at?: Date;
  updated_at?: Date;
}